Binary tree functions
Max height using stack, and queue.
Deleting element in a Binary tree.

Pointer Confution

This commit includes a short code to clarify how pointers are passed in java and how changing pointers in functions effects the actual variables in calling function. Generally, a copy of pointer pointing to the same location is passed. Changing data in the pointing location changes the data in calling function, but changing pointer passed will not reflect on the pointer in calling function.